NGS-general: relocate 'reorder_fasta.py' into bcftbx.cli.
