+++ bin/reorder_fasta.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+#
+#     reorder_fasta.py: reorder chromosomes in a FASTA file
+#     Copyright (C) University of Manchester 2021 Peter Briggs
+#
+from bcftbx.cli.reorder_fasta import main
+if __name__ == "__main__":
+     main()
